+++ src/components/loader/Loader.test.tsx
@@ -1,8 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
 import { Loader } from "./Loader";
-import { LoaderSize } from "../../types/Size";
-import { LoaderColor } from "../../types/LoaderColors";
 
 describe("Loader component", () => {
 	it("renders Loader component with default props", () => {
@@ -19,12 +17,7 @@ describe("Loader component", () => {
 	});
 
 	it("renders Loader component with size xs", () => {
-		render(
-			<Loader
-				size={LoaderSize.EXTRASMALL}
-				dataAttributes={{ "data-testid": "loader" }}
-			/>
-		);
+		render(<Loader size="xs" dataAttributes={{ "data-testid": "loader" }} />);
 		const loader = screen.getByTestId("loader");
 		expect(loader).toHaveClass("w-[22px]");
 		expect(loader).toHaveClass("h-[22px]");
@@ -32,12 +25,7 @@ describe("Loader component", () => {
 	});
 
 	it("renders Loader component with size sm", () => {
-		render(
-			<Loader
-				size={LoaderSize.SMALL}
-				dataAttributes={{ "data-testid": "loader" }}
-			/>
-		);
+		render(<Loader size="sm" dataAttributes={{ "data-testid": "loader" }} />);
 		const loader = screen.getByTestId("loader");
 		expect(loader).toHaveClass("w-[42px]");
 		expect(loader).toHaveClass("h-[42px]");
@@ -45,12 +33,7 @@ describe("Loader component", () => {
 	});
 
 	it("renders Loader component with size lg", () => {
-		render(
-			<Loader
-				size={LoaderSize.LARGE}
-				dataAttributes={{ "data-testid": "loader" }}
-			/>
-		);
+		render(<Loader size="lg" dataAttributes={{ "data-testid": "loader" }} />);
 		const loader = screen.getByTestId("loader");
 		expect(loader).toHaveClass("w-[140px]");
 		expect(loader).toHaveClass("h-[140px]");
@@ -59,10 +42,7 @@ describe("Loader component", () => {
 
 	it("renders Loader component with color success", () => {
 		render(
-			<Loader
-				color={LoaderColor.SUCCESS}
-				dataAttributes={{ "data-testid": "loader" }}
-			/>
+			<Loader color="success" dataAttributes={{ "data-testid": "loader" }} />
 		);
 		const loader = screen.getByTestId("loader");
 		expect(loader).toHaveClass("border-success-500");
@@ -70,10 +50,7 @@ describe("Loader component", () => {
 
 	it("renders Loader component with color danger", () => {
 		render(
-			<Loader
-				color={LoaderColor.DANGER}
-				dataAttributes={{ "data-testid": "loader" }}
-			/>
+			<Loader color="danger" dataAttributes={{ "data-testid": "loader" }} />
 		);
 		const loader = screen.getByTestId("loader");
 		expect(loader).toHaveClass("border-danger-500");