+++ src/types/Size.ts
@@ -1,12 +1,5 @@
 export enum Size {
-	SMALL = 'small',
-	MEDIUM = 'medium',
-	LARGE = 'large'
-};
-
-export enum LoaderSize {
-	EXTRASMALL = "xs",
-	SMALL = "sm",
-	MEDIUM = "m",
-	LARGE = "lg",
+	SMALL = "small",
+	MEDIUM = "medium",
+	LARGE = "large",
 }