+++ stories/components/Chart.stories.tsx
@@ -283,9 +283,7 @@ export const AreaChart = {
 				type: "gradient",
 				gradient: {
 					shadeIntensity: 1,
-					opacityFrom: 0.7,
-					opacityTo: 0.9,
-					stops: [0, 90, 100],
+					gradientToColors: ["#0ea5e9"],
 				},
 			},
 			colors: ["#0ea5e9"],